+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symbols/internal/JAccessibleDeclarationSymbol.java
@@ -4,13 +4,7 @@
 
 package net.sourceforge.pmd.lang.java.symbols.internal;
 
-import java.lang.reflect.Modifier;
-import java.util.Objects;
-import java.util.Optional;
-
 import net.sourceforge.pmd.lang.ast.Node;
-import net.sourceforge.pmd.lang.java.ast.ASTAnyTypeDeclaration;
-import net.sourceforge.pmd.lang.java.ast.AccessNode;
 
 
 /**
@@ -20,31 +14,11 @@ import net.sourceforge.pmd.lang.java.ast.AccessNode;
  * @author Clément Fournier
  * @since 7.0.0
  */
-public abstract class JAccessibleDeclarationSymbol<N extends Node>
-    extends AbstractDeclarationSymbol<N> {
+public interface JAccessibleDeclarationSymbol<N extends Node> extends JDeclarationSymbol<N> {
 
-    protected final int myModifiers;
-    private final JClassSymbol myEnclosingClass;
-
-
-    JAccessibleDeclarationSymbol(int modifiers, String simpleName, Class<?> enclosingClass) {
-        super(simpleName);
-        this.myModifiers = modifiers;
-        this.myEnclosingClass = enclosingClass == null ? null
-                                                       : JClassSymbol.create(enclosingClass);
-    }
-
-
-    JAccessibleDeclarationSymbol(N node, int modifiers, String simpleName) {
-        super(node, simpleName);
-        this.myModifiers = modifiers;
-        this.myEnclosingClass =
-            // doesn't handle anonymous type declaration until we handle #905
-            Optional.ofNullable(node.getFirstParentOfType(ASTAnyTypeDeclaration.class))
-                    .map(JClassSymbol::create)
-                    .orElse(null);
-    }
-
+    // TODO #905 is an enormous hole is this architecture, because
+    // one cannot build a JClassSymbol from an anonymous class because
+    // it's not an ASTAnyTypeDeclaration
 
     /**
      * Returns the class that directly encloses this declaration.
@@ -54,119 +28,19 @@ public abstract class JAccessibleDeclarationSymbol<N extends Node>
      * This is necessarily an already resolved symbol, because
      * 1. if it's obtained from reflection, then the enclosing class is available, in all likelyhood
      * 2. if it's obtained from an AST, then the enclosing class is in the same source file so we can
-     *    know about it
+     * know about ita
      */
-    public JClassSymbol getEnclosingClass() {
-        return myEnclosingClass;
-    }
+    JClassSymbol getEnclosingClass();
 
 
-    public final boolean isPublic() {
-        return Modifier.isPublic(myModifiers);
-    }
+    boolean isPublic();
 
 
-    public final boolean isPrivate() {
-        return Modifier.isPrivate(myModifiers);
-    }
+    boolean isPrivate();
 
 
-    public final boolean isProtected() {
-        return Modifier.isProtected(myModifiers);
-    }
+    boolean isProtected();
 
 
-    public final boolean isPackagePrivate() {
-        return (myModifiers & (Modifier.PRIVATE | Modifier.PROTECTED | Modifier.PUBLIC)) == 0;
-    }
-
-
-    static int accessNodeToModifiers(AccessNode accessNode) {
-
-        /*
-        AccessNode:                     java.lang.reflect.Modifier
-
-        int PUBLIC = 0x0001;            0x00000001;
-        int PROTECTED = 0x0002;         0x00000004;
-        int PRIVATE = 0x0004;           0x00000002;
-        int ABSTRACT = 0x0008;          0x00000400;
-        int STATIC = 0x0010;            0x00000008;
-        int FINAL = 0x0020;             0x00000010;
-        int SYNCHRONIZED = 0x0040;      0x00000020;
-        int NATIVE = 0x0080;            0x00000100;
-        int TRANSIENT = 0x0100;         0x00000080;
-        int VOLATILE = 0x0200;          0x00000040;
-        int STRICTFP = 0x1000;          0x00000800;
-        int DEFAULT = 0x2000;           -----------
-
-        */
-
-        // TODO AccessNode should use the conventions of the standard reflection API
-
-        int result = 0;
-        if (accessNode.isPublic()) {
-            result |= Modifier.PUBLIC;
-        }
-        if (accessNode.isPrivate()) {
-            result |= Modifier.PRIVATE;
-        }
-        if (accessNode.isProtected()) {
-            result |= Modifier.PROTECTED;
-        }
-        if (accessNode.isStatic()) {
-            result |= Modifier.STATIC;
-        }
-        if (accessNode.isFinal()) {
-            result |= Modifier.FINAL;
-        }
-        if (accessNode.isSynchronized()) {
-            result |= Modifier.SYNCHRONIZED;
-        }
-        if (accessNode.isVolatile()) {
-            result |= Modifier.VOLATILE;
-        }
-        if (accessNode.isTransient()) {
-            result |= Modifier.TRANSIENT;
-        }
-        if (accessNode.isNative()) {
-            result |= Modifier.NATIVE;
-        }
-        if (accessNode.isAbstract()) {
-            result |= Modifier.ABSTRACT;
-        }
-        if (accessNode.isStrictfp()) {
-            result |= Modifier.STRICT;
-        }
-
-        return result;
-    }
-
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        if (!super.equals(o)) {
-            return false;
-        }
-        JAccessibleDeclarationSymbol<?> that = (JAccessibleDeclarationSymbol<?>) o;
-        return myModifiers == that.myModifiers && myEnclosingClass.equals(that.myEnclosingClass);
-    }
-
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(super.hashCode(), myModifiers);
-    }
-
-
-    protected static JClassSymbol toResolvable(Class<?> clazz) {
-        return JClassSymbol.create(clazz);
-    }
-
-
+    boolean isPackagePrivate();
 }
