+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symbols/internal/JSimpleTypeDeclarationSymbol.java
@@ -9,10 +9,11 @@ import net.sourceforge.pmd.lang.ast.Node;
 
 /**
  * A reference type that can be referred to using a simple name.
- * These include references to {@linkplain JResolvableClassSymbol class or interfaces}
+ * These include references to class or interfaces (be they
+ * {@linkplain JClassSymbol resolved} or only {@link JResolvableClassSymbol symbolic}),
  * and references to {@linkplain JTypeParameterSymbol type parameters},
- * but not array types or parameterized types. Primitive types are excluded
- * as well because that wouldn't be useful.
+ * but not array types or parameterized types. Primitive types are
+ * excluded as well because that wouldn't be useful.
  *
  * @param <N> Type of AST node that can represent this type of declaration
  *
@@ -21,11 +22,21 @@ import net.sourceforge.pmd.lang.ast.Node;
  */
 public interface JSimpleTypeDeclarationSymbol<N extends Node> extends JDeclarationSymbol<N> {
 
+
+    /**
+     * Returns true if this is a resolved class reference, in
+     * which case it can be safely downcast to {@link JClassSymbol}.
+     */
+    default boolean isResolvedClass() {
+        return this instanceof JClassSymbol;
+    }
+
+
     /**
      * Returns true if this is a class reference, in
      * which case it can be safely downcast to {@link JResolvableClassSymbol}.
      */
-    default boolean isSymbolicClass() {
+    default boolean isUnresolvedClass() {
         return this instanceof JResolvableClassSymbol;
     }
