+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symbols/internal/JTypeParameterOwnerSymbol.java
@@ -6,10 +6,13 @@ package net.sourceforge.pmd.lang.java.symbols.internal;
 
 import java.util.List;
 
+import net.sourceforge.pmd.lang.java.symbols.internal.impl.JClassSymbolImpl;
+import net.sourceforge.pmd.lang.java.symbols.internal.impl.JMethodSymbolImpl;
+
 
 /**
  * Represents a declaration that can declare type parameters,
- * i.e. {@link JClassSymbol} or {@link JMethodSymbol}.
+ * i.e. {@link JClassSymbolImpl} or {@link JMethodSymbolImpl}.
  *
  * @author Clément Fournier
  * @since 7.0.0
