+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symbols/internal/impl/AbstractDeclarationSymbol.java
@@ -2,12 +2,13 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.java.symbols.internal;
+package net.sourceforge.pmd.lang.java.symbols.internal.impl;
 
 import java.util.Objects;
 import java.util.Optional;
 
 import net.sourceforge.pmd.lang.ast.Node;
+import net.sourceforge.pmd.lang.java.symbols.internal.JDeclarationSymbol;
 
 
 /**
