+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symbols/internal/impl/Lazy.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.java.symbols.internal;
+package net.sourceforge.pmd.lang.java.symbols.internal.impl;
 
 import java.util.Objects;
 import java.util.function.Supplier;
